server/requirements/development.txt
django-cors-headers
isort
pylama
pytest==3.0.*
pytest-django==3.1.*
